refactored dialogs
