+++ Tinycast/Core/Theme.swift
@@ -25,8 +25,8 @@ enum Theme {
         /// Hover highlight behind a popover menu row.
         static let menuRow: CGFloat = 10
         static let menuPanel: CGFloat = 16
-        /// Tinycast's own modal / HUD surface, sized between `menuPanel` and `panel`, so a dialog reads as a smaller sibling of the palette rather than a second palette.
-        static let modal: CGFloat = 20
+        /// Tinycast's own dialog / HUD surface, sized between `menuPanel` and `panel`, so a dialog reads as a smaller sibling of the palette rather than a second palette.
+        static let dialog: CGFloat = 20
         static let thumbnail: CGFloat = 6
         static let card: CGFloat = 10
         static let keyCap: CGFloat = 6
@@ -71,10 +71,10 @@ enum Theme {
         /// Confirmation HUD: it sizes to its message, up to this ceiling, and sits this far above the bottom of the screen.
         static let hudMaxWidth: CGFloat = 420
         static let hudEdgeOffset: CGFloat = 48
-        /// Tinycast's own modal: fixed width, height measured from the SwiftUI content.
-        static let modalWidth: CGFloat = 420
-        /// Leading glyph on a modal, larger than a row icon because it carries the dialog's tone (warning / question).
-        static let modalIcon: CGFloat = 32
+        /// Tinycast's own dialog: fixed width, height measured from the SwiftUI content.
+        static let dialogWidth: CGFloat = 420
+        /// Leading glyph on a dialog, larger than a row icon because it carries the subject the dialog is about.
+        static let dialogIcon: CGFloat = 32
         /// Transient volume HUD shown after any volume or mute command.
         static let hudWidth: CGFloat = 200
         static let hudHeight: CGFloat = 92
@@ -86,6 +86,9 @@ enum Theme {
     enum Duration {
         /// How long the confirmation HUD stays on screen.
         static let hud: TimeInterval = 1.6
+        /// Dialog entrance and exit. The exit is shorter so a confirmed action doesn't feel held up.
+        static let dialogOpen: TimeInterval = 0.18
+        static let dialogClose: TimeInterval = 0.12
         /// Fade-in/out for a hover `Tooltip`.
         static let tooltip: TimeInterval = 0.15
     }
@@ -128,9 +131,9 @@ enum Theme {
         static let glassFrost = Color.white.opacity(0.05)
         /// The violet of the app mark, used only to tint the About support callout.
         static let brand = Color(red: 0.525, green: 0.231, blue: 1.0)
-        /// Destructive tint: a destructive button's label, and a modal's leading glyph for `.warning`/`.error` kind dialogs — both read equally severe; only the default icon's shape tells them apart.
+        /// Destructive tint: a destructive button's label, and the leading glyph of a `.danger` dialog.
         static let destructive = Color.red
-        /// Success tint: a modal's leading glyph for a `.success` kind dialog.
+        /// Success tint: the leading glyph of a `.success` dialog.
         static let success = Color.green
     }
 }
