+++ docs/custom-commands.md
@@ -75,8 +75,10 @@ is dropped while the actual error survives.
 `AppCore.runCustomCommand(id:)` is the one funnel both palette activation and the global hotkey reach,
 so the gate lives there and neither path can bypass it. The palette hides before the dialog it is a
 floating panel and would sit above it. The dialog shows the command text as well as its name; ↵ runs
-it and Escape cancels, with Cancel rendered on the left of the two buttons. The gate is Tinycast's own
-modal, not an `NSAlert` ([ui.md](ui.md#modals--hud)): presentation is `async` with no nested run loop,
+it and Escape cancels, with Cancel rendered on the left of the two buttons. It carries the `terminal`
+glyph the command's launcher row uses, and reads neutral rather than destructive — running a command the
+user wrote themselves wants a deliberate second tap, not a red alarm. The gate is Tinycast's own
+dialog, not an `NSAlert` ([ui.md](ui.md#dialogs--hud)): presentation is `async` with no nested run loop,
 and the presenter itself refuses a second dialog while one is up, so a held shortcut can't stack them.
 
 ### Reporting
