+++ docs/snippets.md
@@ -173,8 +173,8 @@ The confirmation is per snippet and off by default: the only gate is `show_confi
 from the snippet's editor in **Settings → Snippets**. Nothing about it reaches settings backups.
 The feature switch — which carries keyword-monitoring consent — is likewise excluded from backups.
 
-`HUDWindowController` is shared rather than snippet-specific. It takes a message and a `ModalKind`
-(defaulting to `.success`), the same kind vocabulary `ModalWindowController`'s dialogs use, so a
+`HUDWindowController` is shared rather than snippet-specific. It takes a message and a `DialogTone`
+(defaulting to `.success`), the same tone vocabulary `DialogController`'s dialogs use, so a
 custom command confirms a run through the same panel and the same tint rules; system commands'
 success/info feedback uses it too (see [launcher.md](launcher.md#system-commands)). Its leading
 status dot, not an icon, carries the tint. Its capsule uses `Theme.frosted(in:)`, the same
